Sprint 16 MVVM 1.22 Cleared project. PlayerActivity work Search Activity is worked. Before moving SettingsActivity
