+++ app/src/main/AndroidManifest.xml
@@ -29,14 +29,14 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".mediateka.MediaActivity"
+            android:name=".mediateka.ui.MediaActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".main.MainActivity"
+            android:name=".main.ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
