+++ app/src/main/java/com/practicum/playlistmaker/main/ui/MainActivity.kt
@@ -1,14 +1,12 @@
-package com.practicum.playlistmaker.main
+package com.practicum.playlistmaker.main.ui
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.View
 import android.widget.Button
-import androidx.activity.ComponentActivity
 import com.practicum.playlistmaker.R
 import com.practicum.playlistmaker.databinding.ActivityMainBinding
-import com.practicum.playlistmaker.mediateka.MediaActivity
+import com.practicum.playlistmaker.mediateka.ui.MediaActivity
 import com.practicum.playlistmaker.search.ui.SearchActivity
 import com.practicum.playlistmaker.settings.SettingsActivity
