+++ app/src/main/res/layout/activity_media.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".mediateka.MediaActivity">
+    tools:context=".mediateka.ui.MediaActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
